% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan.srm-package.R
\docType{package}
\name{lavaan.srm-package}
\alias{lavaan.srm-package}
\title{The 'lavaan.srm' package.}
\description{
This package provides a convenient interface for fitting a multivariate
social relations model (SRM; Nestler, 2018) using Markov chain Monte Carlo
(MCMC) estimation with the \pkg{rstan} package. This package also provides a
convenient interface for fitting a structural equation model (SEM) to the
SRM results, thus enabling a two-stage estimator of the SR-SEM (Nestler
et al., 2020, 2022) via the \pkg{lavaan} package (Rosseel, 2012).
}
\references{
Nestler, S. (2018). Likelihood estimation of the multivariate social
relations model. \emph{Journal of Educational and Behavioral Statistics, 43}(4),
387--406. \url{https://doi.org/10.3102/1076998617741106}

Nestler, S., Lüdtke, O., & Robitzsch, A. (2020). Maximum likelihood
estimation of a social relations structural equation model.
\emph{Psychometrika, 85}(4), 870--889.
\url{https://doi.org/10.1007/s11336-020-09728-z}

Nestler, S., Lüdtke, O., & Robitzsch, A. (2022). Analyzing longitudinal
social relations model data using the social relations structural equation
model. \emph{Journal of Educational and Behavioral Statistics, 47}(2), 231--260.
\url{https://doi.org/10.3102/10769986211056541}

Rosseel, Y. (2012). \code{lavaan}: An R package for structural equation modeling.
\emph{Journal of Statistical Software, 48}(2), 1--36.
\url{https://doi.org/10.18637/jss.v048.i02}

Stan Development Team. (2022). \code{rstan}: The R interface to Stan.
R package version 2.21.7. https://mc-stan.org
}
