% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan.R
\name{lavaan.srm}
\alias{lavaan.srm}
\title{Fit a \code{lavaan} Model to Multivariate SRM Results}
\usage{
lavaan.srm(model, data, point = "mean", ...)
}
\arguments{
\item{model}{\code{character} specifying an SEM.  See \code{\link[lavaan:model.syntax]{lavaan::model.syntax()}}.}

\item{data}{An object of class \code{\linkS4class{mvSRM}}}

\item{point}{\code{character} indicating choice of posterior point estimate.
Must be \verb{\%in\% c("mean", "median", "mode")}.}

\item{...}{Arguments and \code{\link[lavaan:lavOptions]{lavaan::lavOptions()}} passed to \code{\link[lavaan:lavaan]{lavaan::lavaan()}}}
}
\value{
An object of \code{class?lavaan}.
}
\description{
Estimate a structural equation model (SEM) parameters, using MCMC-estimated
summary statistics from a multivariate social relations model (mvSRM;
Nestler, \href{https://doi.org/10.3102/1076998617741106}{2018}) as data.
}
