% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvSRM-class.R
\docType{class}
\name{mvSRM-class}
\alias{mvSRM-class}
\alias{show,mvSRM-method}
\alias{summary,mvSRM-method}
\alias{as.matrix,mvSRM-method}
\alias{vcov,mvSRM-method}
\title{Class for a Multivariate SRM using Stan}
\usage{
\S4method{show}{mvSRM}(object)

\S4method{summary}{mvSRM}(
  object,
  component = c("case", "dyad"),
  stat = c("sd", "cor"),
  point = "mean",
  interval = "central",
  credMass = 0.95,
  as.stanfit = FALSE,
  ...
)

\S4method{as.matrix}{mvSRM}(x, stat, component, point = "mean", ...)

\S4method{vcov}{mvSRM}(
  object,
  component,
  meanstructure = FALSE,
  keep,
  drop,
  add.names.attr = FALSE,
  ...
)
}
\arguments{
\item{object, x}{An object of class \code{mvSRM}}

\item{component}{\code{character} specifying which SRM component to report
estimated parameters for. Can be \verb{\%in\% c("case", "dyad")}, as well as
\code{"group"} when group effects are modeled.  Ignored when \code{stat="mean"}
because means are only a group-level statistic.
Multiple options can be passed to \code{summary()} method, but \code{as.matrix()}
only uses the first.}

\item{stat}{\code{character} specifying which summary statistic to report for
\code{component}. Can be \verb{\%in\% c("sd", "cor", "cov")}, as well as \code{"mean"} when
\code{fixed.groups=FALSE}. The \code{summary()} method will always return the
means as a group-level statistic, even for a single round-robin group.}

\item{point}{For \code{as.matrix()}, a length-1 \code{character} or \code{numeric}
indicating which posterior summary statistic should be returned.  Can be
\verb{\%in\% c("mean", "median", "mode", "min", "max", "hdi")}.
\itemize{
\item \code{point="mode"} calls \code{\link[modeest:mlv]{modeest::mlv()}}, whose additional arguments can be
passed via \code{...} (recommended to avoid warnings about default \verb{method=}).
\item \code{point="hdi"} calls \code{\link[HDInterval:hdi]{HDInterval::hdi()}}, whose additional arguments can
be passed via \code{...} (e.g., to overwrite the default \code{credMass=.95}).
Unlike any other \verb{point=} options for \code{as.matrix()}, \code{"hdi"} returns two
values per parameter: the lower and upper credible-interval limits.
This affects how the output is formatted.
\item \code{point=numeric(1)} between 0 and 1 (inclusive) is accepted by
\code{as.matrix()}, passed to \code{\link[stats:quantile]{stats::quantile()}} as the \verb{probs=} argument;
other arguments can be passed via \code{...}.
\item For the \code{summary()} method, multiple options can be passed, but must be
\verb{\%in\% c("mean", "median", "mode")} or set to \code{NULL} to avoid returning
point estimates.
}}

\item{interval}{\code{character} indicating the type of uncertainty/credible
interval estimate.  Can be \verb{\%in\% c("central", "hdi")} or set to
\code{NULL} to avoid returning interval estimates.}

\item{credMass}{\code{numeric} passed to \code{\link[HDInterval:hdi]{HDInterval::hdi()}}, or used to request
\verb{probs=} from \code{\link[stats:quantile]{stats::quantile()}} when \code{interval="central"}.}

\item{as.stanfit}{\code{logical} indicating whether to use the \code{summary()} method
defined for a \code{\linkS4class{stanfit}} object.  Useful for obtaining
diagnostics (e.g., R-hat, effective sample size) about MCMC results.  If
\code{TRUE}, further arguments to \link[rstan]{summary,stanfit-method} can be
passed via \code{...}}

\item{...}{Additional arguments passed to or from other methods}

\item{meanstructure}{\code{logical} indicating whether the \code{vcov()} method
includes posterior variability of the mean estimates (only when
\code{component="group"}).}

\item{keep, drop}{Variables to include (\verb{keep=}) or exclude (\verb{drop=}) from
the result.  \verb{drop=} is ignored when \verb{keep=} is specified.  For case- or
dyad-level variables, \code{keep}/\code{drop} can be any combination of the original
names of round-robin variables passed to \code{\link[=mvsrm]{mvsrm()}} or the names of
specific components, which include \code{c("out","in")} suffixes for case-level
components or \code{c("ij","ji")} suffixes for dyad/relationship-level
components.}
}
\value{
\itemize{
\item \code{show()} simply prints a message stating the class and inheritance.
\item \code{as.matrix()} returns a vector or matrix of the requested posterior
summary for the specified SRM parameter and level of analysis.
\item \code{summary()} loops over \code{as.matrix()} to return a \code{list()} of potentially
multiple summaries of potentially multiple SRM parameters and components.
\item \code{vcov()} provides the posterior covariance matrix of sampled (co)variance
parameters at the requested level of analysis.  This is only useful for
the 2-stage estimation provided using \code{\link[lavaan:lavaan]{lavaan::lavaan()}}.
}
}
\description{
This class extends the \code{\linkS4class{stanfit}} class, created by
fitting a multivariate social relations model (mvSRM; Nestler,
\href{https://doi.org/10.3102/1076998617741106}{2018}) using the Stan
(Carpenter et al., \href{https://doi.org/10.18637/jss.v076.i01}{2017}) software.
}
\section{Slots}{

\describe{
\item{\code{call}}{A copy of the \code{call} to the \code{\link[=mvsrm]{mvsrm()}} function that generated
the object.}

\item{\code{nobs}}{A vector of group-level (when relevant), case-level, and
dyad-level sample sizes.}

\item{\code{IDs}}{\code{list} of \code{data.frame}s containing ID variables from data sets at
each level of analysis.  Stored as integers with \code{\link[=attributes]{attributes()}} that are
the original labels in the user's \verb{data=} (and optionally \verb{case_data=}
and \verb{group_data=}) passed to \code{\link[=mvsrm]{mvsrm()}}.}

\item{\code{varNames}}{\code{list} of variable names at each level of analysis.  This is
used to ease the writing of methods to extract information by variable,
since each variable has 2 person-level and 2 dyad-level components.}

\item{\code{parNames}}{\code{list} of parameters, organized in sets:
\itemize{
\item \verb{$mu} contains \code{"Mvec"} (the mean vector, if estimated)
\item \verb{$sigma} contains \code{c("s_rr","S_p")} (also \code{"S_g"}, if estimated). These
are the \emph{SD}s of relationship-level and case-level components of
round-robin variables (and \emph{SD}s of covariates at those levels), as well
as group-level components when relevant.
\item \verb{$corr} contains \code{c("Rd2","Rp")} (also \code{"Rg"}, if estimated). These are
correlation matrices among relationship-level and case-level components
of round-robin variables (and correlations with covariates at those
levels), as well as group-level components when relevant.
\item \verb{$derived} contains \code{"Rsq"}: a matrix with 3-4 columns and 1 row per
round-robin variable listed in \code{object@varNames$RR}. When group-level
variances are estimated, the first of 4 columns contains the proportion
of variance (in each round-robin variable) accounted for by group-mean
differences.  The remaining columns are proportions of variance
accounted for by outgoing (e.g., actor, perceiver, ego) effects, by
incoming (e.g., partner, target, alter) effects, and by relationship
effects (confounded with any dyad-level measurement error).
\item \verb{$components} contains the relationship-level and case-level components
themselves when \code{object@call$saveComp=TRUE}. This makes the object take
much more computer memory, but it would enable the plausible-values
approach described by Lüdtke et al. (2018).
}}
}}

\section{Objects from the Class}{
 See the \code{\link[=mvsrm]{mvsrm()}} function for details.
}

\references{
Carpenter, B., Gelman, A., Hoffman, M. D., Lee, D., Goodrich, B.,
Betancourt, M., ... & Riddell, A. (2017).
Stan: A probabilistic programming language.
\emph{Journal of Statistical Software, 76}(1).
\url{https://doi.org/10.18637/jss.v076.i01}

Lüdtke, O., Robitzsch, A., & Trautwein, U. (2018). Integrating covariates
into social relations models: A plausible values approach for handling
measurement error in perceiver and target effects.
\emph{Multivariate Behavioral Research, 53}(1), 102--124.
\url{https://doi.org/10.1080/00273171.2017.140679}

Nestler, S. (2018). Likelihood estimation of the multivariate social
relations model. \emph{Journal of Educational and Behavioral Statistics, 43}(4),
387--406. \url{https://doi.org/10.3102/1076998617741106}
}
\seealso{
\code{\link[=mvsrm]{mvsrm()}}
}
\author{
Terrence D. Jorgensen (University of Amsterdam;
\email{TJorgensen314@gmail.com})
}
