% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{srm_priors}
\alias{srm_priors}
\title{Default priors for multivariate SRM}
\usage{
srm_priors(
  data,
  group_data,
  case_data,
  modelG = FALSE,
  modelM = FALSE,
  SDby = "sd",
  decomp_rr = c(dyad = 0.5, out = 0.2, `in` = 0.2, group = 0.1),
  decomp_c = c(case = 0.9, group = 0.1)
)
}
\arguments{
\item{data, group_data, case_data}{See the \code{\link[=mvsrm]{mvsrm()}} function for details.
However, the data sets should contain \strong{only} modeled variables (no
ID variables), and \verb{data=} must contain round-robin variables in long
format (one column per variable, 2 rows per dyad).}

\item{modelG}{\code{logical} indicating whether to estimate group-level (co)variances}

\item{modelM}{\code{logical} indicating whether to estimate means. Setting \code{TRUE}
presupposes either that there is only 1 round-robin group
(\code{modelG=FALSE}) or that group-level effects will not be partialed
out as fixed effects (\code{fixed.groups=FALSE} when \code{\link[=mvsrm]{mvsrm()}} is called).}

\item{SDby}{\code{character} indicating how to choose priors for \emph{SD}s of each
round-robin variable component (i.e., dyad- and case-level, and
potentially group-level if \code{modelG=TRUE}). Options include
\code{"sd"} to base it on the (total) sample \emph{SD}, or \code{"range"} to
base it on the empirical minimum and maximum observation. The latter
is likely to overestimate the actual \emph{SD}, and it might only be a
sensible heuristic when using Likert-type data, due to their
limited response options that determine the maximum possible \emph{SD}.}

\item{decomp_rr}{\code{numeric} vector of weights (internally rescaled so sum ==
1) indicating hypothesized proportions of total variance attributable
to each round-robin component. The vector should therefore have the
following \code{\link[base:names]{base::names()}}: \code{c("dyad","out","in")}, as well as
\code{"group"} when \code{modelG=TRUE}. Note that the name \code{"in"} must be in
backticks when assigning names within the \code{\link[base:c]{base::c()}} function (see
the default value in \strong{Usage}), in order to avoid confusion with the
\code{\link[base:Reserved]{base::Reserved()}} word in R.  The \code{"out"} and \code{"in"} components
correspond to what are often called "perceivers" and "targets" in
SRMs of interpersonal perception, or "actors" and "partners" in SRMs
of group behavior. The same hypothesized decomposition is applied to
each round-robin variable, unless the user provides a list of such
vectors (with length equal to the number of variables in \verb{data=}) to
apply to each round-robin variable. A list of vectors can have names
that correspond to those in \verb{data=}, but will otherwise be assumed to
follow the same order.}

\item{decomp_c}{\code{numeric} vector (or list of vectors) similar to \code{decomp_rr}
argument, but for the decomposition of \code{case_data} into case- and
group-level variance components.  Ignored unless \code{modelG=TRUE}.}
}
\value{
A \code{list} of hyperparameters, with an \code{attr(,"Nvars")} indicating
the number of (components of) variables modeled at each level (e.g.,
the \code{case} level includes 2 components per round-robin variable
plus the number of variables in \verb{case_data=}).
}
\description{
Priors used during MCMC estimation of a multivariate social relations model
(mvSRM; Nestler, \href{https://doi.org/10.3102/1076998617741106}{2018}) using the
Stan (Carpenter et al., \href{https://doi.org/10.18637/jss.v076.i01}{2017}) software.
}
\examples{
## use example data from the ?srm::srm help page
data(data.srm01, package = "srm")

## ignoring group-level variance, not estimating means
## (i.e., mvsrm() called with fixed.groups = TRUE)
srm_priors(data.srm01[5:7])

## include group-level variance and means (fixed.groups = FALSE)
srm_priors(data.srm01[5:7], modelG = TRUE, modelM = TRUE)

## base SD hyperparameters on observed range instead of observed SD
srm_priors(data.srm01[5:7], SDby = "range")

}
\references{
Carpenter, B., Gelman, A., Hoffman, M. D., Lee, D., Goodrich, B.,
Betancourt, M., Brubaker, M., Guo, J., Li, P., & Riddell, A. (2017).
Stan: A probabilistic programming language.
\emph{Journal of Statistical Software, 76}(1).
\url{https://doi.org/10.18637/jss.v076.i01}

Nestler, S. (2018). Likelihood estimation of the multivariate social
relations model. \emph{Journal of Educational and Behavioral Statistics, 43}(4),
387--406. \url{https://doi.org/10.3102/1076998617741106}
}
